Move ui specific modules in package gui
