+++ arelle/ModelXbrl.py
@@ -13,6 +13,8 @@ from arelle.ModelObject import ModelObject
 from arelle.Locale import format_string
 from arelle.PrototypeInstanceObject import FactPrototype, DimValuePrototype
 from arelle.ValidateXbrlDimensions import isFactDimensionallyValid
+from arelle.gui import ViewWinDTS
+
 ModelRelationshipSet = None # dynamic import
 
 profileStatNumber = 0
@@ -396,7 +398,6 @@ class ModelXbrl:
         self.closeFileSource= True
         del self.entryLoadingUrl
         # reload dts views
-        from arelle import ViewWinDTS
         for view in self.views:
             if isinstance(view, ViewWinDTS.ViewDTS):
                 self.modelManager.cntlr.uiThreadQueue.put((view.view, []))
