+++ arelle/gui/CntlrQuickBooks.py
