+++ arelle/gui/CntlrWinTooltip.py
