+++ arelle/gui/DialogAbout.py
