+++ arelle/gui/DialogArcroleGroup.py
@@ -10,8 +10,8 @@ try:
 except ImportError:
     from ttk import Frame, Button
 import os, re
-from arelle.UiUtil import gridHdr, gridCell, gridCombobox, label, checkbox
-from arelle.CntlrWinTooltip import ToolTip
+from arelle.gui.UiUtil import   gridCombobox, label, checkbox
+from arelle.gui.CntlrWinTooltip import ToolTip
 from arelle import XbrlConst
 
 '''
