+++ arelle/gui/DialogFind.py
@@ -10,8 +10,8 @@ try:
 except ImportError:
     from ttk import Frame, Button
 import re
-from arelle.UiUtil import gridHdr, gridCell, gridCombobox, label, checkbox, radiobutton
-from arelle.CntlrWinTooltip import ToolTip
+from arelle.gui.UiUtil import gridHdr, gridCell, gridCombobox, label, checkbox, radiobutton
+from arelle.gui.CntlrWinTooltip import ToolTip
 from arelle import ModelDocument, XPathContext, XPathParser, XmlUtil
 from arelle.ModelDtsObject import ModelConcept
 from arelle.ModelInstanceObject import ModelFact
