+++ arelle/gui/DialogOpenTaxonomyPackage.py
