+++ arelle/gui/DialogPluginManager.py
@@ -9,12 +9,14 @@ based on pull request 4
 '''
 from tkinter import Toplevel, font, messagebox, VERTICAL, HORIZONTAL, N, S, E, W
 from tkinter.constants import DISABLED, ACTIVE
+from arelle.gui import DialogURL
+
 try:
     from tkinter.ttk import Treeview, Scrollbar, Frame, Label, Button
 except ImportError:
     from ttk import Treeview, Scrollbar, Frame, Label, Button
-from arelle import PluginManager, DialogURL
-from arelle.CntlrWinTooltip import ToolTip
+from arelle import PluginManager
+from arelle.gui.CntlrWinTooltip import ToolTip
 import re, os, time
 
 def dialogPluginManager(mainWin):
