+++ arelle/gui/UiUtil.py
