+++ arelle/gui/ViewWinDiffs.py
@@ -4,15 +4,13 @@ Created on Mar 21, 2011
 @author: Mark V Systems Limited
 (c) Copyright 2011 Mark V Systems Limited, All rights reserved.
 '''
-from collections import defaultdict
-import os
 from tkinter import *
 try:
     from tkinter.ttk import *
 except ImportError:
     from ttk import *
-from arelle.ViewWinRelationshipSet import viewRelationshipSet
-from arelle.ViewWinConcepts import viewConcepts
+from arelle.gui.ViewWinRelationshipSet import viewRelationshipSet
+from arelle.gui.ViewWinConcepts import viewConcepts
 from arelle import ModelVersObject, XbrlConst
 
 class ViewWinDiffs:
