+++ arelle/gui/ViewWinRssFeed.py
@@ -5,12 +5,12 @@ Created on Mar 09, 2011
 (c) Copyright 2010 Mark V Systems Limited, All rights reserved.
 '''
 from tkinter import *
+from arelle.gui import ViewWinTree
+
 try:
     from tkinter.ttk import *
 except ImportError:
     from ttk import *
-import os
-from arelle import (ViewWinTree, ModelDocument)
 
 def viewRssFeed(modelXbrl, tabWin):
     view = ViewRssFeed(modelXbrl, tabWin)
