+++ arelle/gui/ViewWinTests.py
@@ -4,13 +4,14 @@ Created on Oct 17, 2010
 @author: Mark V Systems Limited
 (c) Copyright 2010 Mark V Systems Limited, All rights reserved.
 '''
-from tkinter import *
+from arelle.gui import ViewWinTree
+
 try:
     from tkinter.ttk import *
 except ImportError:
     from ttk import *
 import os
-from arelle import (ViewWinTree, ModelDocument)
+from arelle import ( ModelDocument)
 
 def viewTests(modelXbrl, tabWin):
     view = ViewTests(modelXbrl, tabWin)
