+++ arelle/gui/__init__.py
@@ -1,0 +1,10 @@
+__author__ = 'Herman Fischer'
+__copyright__ = "Copyright 2010-2012 Mark V Systems Limited, All rights reserved."
+__credits__ = []
+__license__ = "Apache-2"
+__version__ = "0.1"
+__status__ = "Production"
+
+"""
+This package contains the tkinter GUI for Arelle
+"""
