+++ arelleGUI.pyw
@@ -6,6 +6,6 @@ Use this module to start Arelle in windowing interactive UI
 @author: Mark V Systems Limited
 (c) Copyright 2011 Mark V Systems Limited, All rights reserved.
 '''
-from arelle import CntlrWinMain
+from arelle.gui import CntlrWinMain
 
 CntlrWinMain.main()
